Finished the parsing of function types.
Added the function desugar infix notation which remaps infix
identifiers to prefix transformers. This does not seem to be
tractable in the long run but it was a simple fix that seems
flexible.

At this point more unit tests need to be written both testing
that correct syntax parses correctly and incorrect syntax is
ruled out. Since I am running behind I think that I will finish
type checking, cast insertion, and writing an interpreter for
the ast before I revisit testing.